SQL Server Migration

+++ FintrakBanking.Entities/StagingModels/STG_LOAN_WATCHLIST.cs
@@ -36,7 +36,8 @@ namespace FintrakBanking.Entities.StagingModels
         public int? FEEFREQUENCYTYPEID { get; set; }
 
         [Column(TypeName = "float")]
-        public decimal? INTERESTRATE { get; set; }
+        public double? INTERESTRATE { get; set; }
+        //public decimal? INTERESTRATE { get; set; }
 
         public DateTime? EFFECTIVEDATE { get; set; }
